+++ app/Http/Controllers/ProyectoController.php
@@ -4,25 +4,42 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\Database\QueryException;
+use Illuminate\Contracts\Encryption\DecryptException;
+use Illuminate\Database\Eloquent\ModelNotFoundException;
+use Illuminate\Support\Facades\Validator;
 use App\Pais;
 use App\Comuna;
 use App\Provincia;
 use App\Region;
 use App\Proyecto;
+use Session;
 use DB;
-use Illuminate\Database\QueryException;
-use Illuminate\Contracts\Encryption\DecryptException;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
-use Illuminate\Support\Facades\Validator;
 
 class ProyectoController extends Controller
 {
     public function index()
     {
+        if (!Session::has('idUsuario') && !Session::has('idTipoUsuario') && !Session::has('nombre') && !Session::has('apellido') && !Session::has('correo') && !Session::has('rut')) {
+            return abort(401);
+        }
+        if (Session::has('idTipoUsuario')) {
+            if (Session::get('idTipoUsuario') != 3 && Session::get('idTipoUsuario') != 10) {
+                return abort(401);
+            }
+        }
     	return view('admin.proyectos.index');
     }
     public function create()
     {
+        if (!Session::has('idUsuario') && !Session::has('idTipoUsuario') && !Session::has('nombre') && !Session::has('apellido') && !Session::has('correo') && !Session::has('rut')) {
+            return abort(401);
+        }
+        if (Session::has('idTipoUsuario')) {
+            if (Session::get('idTipoUsuario') != 3 && Session::get('idTipoUsuario') != 10) {
+                return abort(401);
+            }
+        }
     	$paises = Pais::pluck('nombrePais','idPais');
         $regiones = Region::pluck('nombreRegion','idRegion');
         $provincias = Provincia::pluck('nombreProvincia','idProvincia');
@@ -33,10 +50,20 @@ class ProyectoController extends Controller
     {
     	try {
             $validator = Validator::make($request->all(), [
-                'fotoPortada' => 'max:102400'
+                'fotoPortada' => 'required|max:102400',
+                'nombreProyecto' => 'required',
+                'idPais' => 'required',
+                'idRegion' => 'required',
+                'idProvincia' => 'required',
+                'idComuna' => 'required',
+                'direccion' => 'required',
+                'numeracionProyecto' => 'required',
+                'codigoPostal' => 'required',
+                'latitud' => 'required',
+                'longitud' => 'required'
             ]);
             if ($validator->fails()) {
-                toastr()->info('El archivo no puede pasar de los 100MB');
+                toastr()->info('El archivo no puede pasar de los 100MB, los datos no puede venir vacios');
                 return back();
             }
             DB::beginTransaction();
@@ -74,6 +101,14 @@ class ProyectoController extends Controller
     }
     public function edit($idProyecto)
     {
+        if (!Session::has('idUsuario') && !Session::has('idTipoUsuario') && !Session::has('nombre') && !Session::has('apellido') && !Session::has('correo') && !Session::has('rut')) {
+            return abort(401);
+        }
+        if (Session::has('idTipoUsuario')) {
+            if (Session::get('idTipoUsuario') != 3 && Session::get('idTipoUsuario') != 10) {
+                return abort(401);
+            }
+        }
     	$proyecto = Proyecto::find($idProyecto);
     	$paises = Pais::pluck('nombrePais','idPais');
         $regiones = Region::pluck('nombreRegion','idRegion');
@@ -85,10 +120,19 @@ class ProyectoController extends Controller
     {
     	try {
             $validator = Validator::make($request->all(), [
-                'fotoPortada' => 'max:102400'
+                'nombreProyecto' => 'required',
+                'idPais' => 'required',
+                'idRegion' => 'required',
+                'idProvincia' => 'required',
+                'idComuna' => 'required',
+                'direccion' => 'required',
+                'numeracionProyecto' => 'required',
+                'codigoPostal' => 'required',
+                'latitud' => 'required',
+                'longitud' => 'required'
             ]);
             if ($validator->fails()) {
-                toastr()->info('El archivo no puede pasar de los 100MB');
+                toastr()->info('Los datos no puede venir vacios');
                 return back();
             }
             DB::beginTransaction();
@@ -100,7 +144,9 @@ class ProyectoController extends Controller
 	            }
 	            $proyecto = Proyecto::find($idProyecto);
                 if ($imgName != null) {
-                    unlink($proyecto->fotoPortada);
+                    if(file_exists($proyecto->fotoPortada)){
+                        unlink($proyecto->fotoPortada);
+                    }
                 }
 	            $proyecto->fill($request->all());
 	            if ($imgName != null) {
@@ -139,7 +185,9 @@ class ProyectoController extends Controller
     			$proyecto = Proyecto::find($idProyecto);
 	            toastr()->success('Eliminado Correctamente', 'El proyecto: '.$proyecto->nombreProyecto.' ha sido eliminado correctamente', ['timeOut' => 9000]);
                 if ($proyecto->fotoPortada) {
-                    unlink($proyecto->fotoPortada);
+                    if(file_exists($proyecto->fotoPortada)){
+                        unlink($proyecto->fotoPortada);
+                    }
                 }
 	            $proyecto->delete();
     		DB::commit();
